+++ quantforge/__init__.py
@@ -46,6 +46,7 @@ from .double_heston import double_heston_price, double_heston_smile
 from .rough_heston import rough_heston_price, rough_heston_smile
 from .double_heston_calib import calibrate_double_heston
 from .heston_calib import calibrate_heston
+from .lsv import calibrate_leverage as calibrate_lsv_leverage
 from .kou import kou_price, kou_smile
 from .cgmy import cgmy_price, cgmy_smile
 from .nig import nig_price, nig_smile
@@ -194,7 +195,7 @@ try:  # pragma: no cover - trivial availability branch
 except ImportError:  # pragma: no cover
     HAS_NUMPY = False
 
-__version__ = "1.130.0"
+__version__ = "1.131.0"
 
 __all__ = [
     "OptionType",
@@ -249,6 +250,7 @@ __all__ = [
     "rough_heston_smile",
     "calibrate_double_heston",
     "calibrate_heston",
+    "calibrate_lsv_leverage",
     "kou_price",
     "kou_smile",
     "cgmy_price",